Adjustments to Stand command
The Stand embed now uses the image as a thumbnail rather than an image on large servers in order to reduce spam.
It also now displays the description of the power.
The colour of the graph in the image now uses the user's name colour for a little bit of flavour, but tweaks it slightly if values are too low or too high to attempt to mitigate some hard to parse colours.