Fixed a major oversight in Meme Cog
Formerly the last meme rolled was simply the ID of the last roll stored in memory.
This caused issues where one server rolling a meme would affect voting and memeinfo commands on another server.
Fixed by changing last_meme_roll to a dict with the guild id as a key, with the ID of the last meme rolled as the value.